ref: move precompiles to separate dir
